+++ CHANGELOG.md
@@ -8,7 +8,7 @@ backwards compatibility.
 
 * LED can be configured to visualize the current position within the loop.
 * Allow reset of the loop position through a control input.
-* Play/pause the delay line using a control input.
+* Pause/resume the delay line using a control input.
 * Allow configuration of the ratio between the tap interval and delay length.
 
 ## 1.2.0